api: move wf::workspace_stream_t into its own header
This allows us to not include opengl.hpp from render_manager.hpp
